+++ pkg/providers/instance/utils/utils.go
@@ -1,0 +1,58 @@
+package utils
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"strings"
+
+	sdkerrors "github.com/Azure/azure-sdk-for-go-extensions/pkg/errors"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice/v9"
+	"github.com/Azure/karpenter-provider-azure/pkg/consts"
+	"github.com/samber/lo"
+	"sigs.k8s.io/controller-runtime/pkg/log"
+)
+
+// HandleProvisioningState inspects the provisioning state of an AKS machine and returns
+func HandleProvisioningState(ctx context.Context, aksMachine *armcontainerservice.Machine) (errorDetails *armcontainerservice.ErrorDetail, pollingErr error, done bool) {
+	provisioningState := lo.FromPtr(aksMachine.Properties.ProvisioningState)
+	switch provisioningState {
+	case consts.ProvisioningStateCreating, consts.ProvisioningStateUpdating:
+		log.FromContext(ctx).V(2).Info("Poller: polling for AKS machine ongoing",
+			"aksMachineID", aksMachine.ID,
+			"provisioningState", provisioningState,
+		)
+		return nil, nil, false
+
+	case consts.ProvisioningStateDeleting:
+		return nil, fmt.Errorf("AKS machine %q sees canceled provisioning state %s", lo.FromPtr(aksMachine.Name), provisioningState), true
+
+	case consts.ProvisioningStateSucceeded:
+		return nil, nil, true
+
+	case consts.ProvisioningStateFailed:
+		if aksMachine.Properties.Status != nil && aksMachine.Properties.Status.ProvisioningError != nil {
+			return aksMachine.Properties.Status.ProvisioningError, nil, true
+		}
+		return nil, fmt.Errorf("AKS machine %q sees fatal provisioning state %s, but ProvisioningError is nil", lo.FromPtr(aksMachine.Name), provisioningState), true
+
+	default:
+		log.FromContext(ctx).V(1).Info("Poller: warning: polling for AKS machine found unrecognized provisioning state, may retry",
+			"aksMachineID", aksMachine.ID,
+			"provisioningState", provisioningState,
+		)
+		return nil, nil, false
+	}
+}
+
+func IsAKSMachineOrMachinesPoolNotFound(err error) bool {
+	if err == nil {
+		return false
+	}
+	azErr := sdkerrors.IsResponseError(err)
+	if azErr != nil && (azErr.StatusCode == http.StatusNotFound || // Covers AKS machines pool not found on PUT machine, GET machine, GET (list) machines, POST agent pool (DELETE machines), and AKS machine not found on GET machine
+		(azErr.StatusCode == http.StatusBadRequest && azErr.ErrorCode == "InvalidParameter" && strings.Contains(azErr.Error(), "Cannot find any valid machines"))) { // Covers AKS machine not found on POST agent pool (DELETE machines)
+		return true
+	}
+	return false
+}
